auto commot

+++ main.cpp
@@ -516,7 +516,7 @@ int getPidByName(char* task_name)
 void tolower(char *str)
 {
     int i=0;
-    for(i = 0; i < sizeof(str); i++)
+    for(i = 0; i <(int) sizeof(str); i++)
     {
            str[i] = tolower(str[i]);
     }
